IAM 848

+++ pkg/roles/service.go
@@ -253,9 +253,12 @@ func (s *Service) DeleteRole(ctx context.Context, ID string) error {
 	// https://go.dev/ref/spec#Send_statements
 	// A send on an unbuffered channel can proceed if a receiver is ready.
 	// A send on a buffered channel can proceed if there is room in the buffer
-	results := make(chan *pool.Result[any], len(s.permissionTypes()))
+	jobs := len(s.permissionTypes()) + 1
+
+	results := make(chan *pool.Result[any], jobs)
 	wg := sync.WaitGroup{}
-	wg.Add(len(s.permissionTypes()))
+	// number of types + 1 for assignees job
+	wg.Add(jobs)
 
 	// TODO @shipperizer use a background operator
 	for _, t := range s.permissionTypes() {
@@ -265,6 +268,12 @@ func (s *Service) DeleteRole(ctx context.Context, ID string) error {
 			&wg,
 		)
 	}
+
+	s.wpool.Submit(
+		s.removeAssigneesFunc(ctx, ID),
+		results,
+		&wg,
+	)
 
 	// wait for tasks to finish
 	wg.Wait()
@@ -326,7 +335,47 @@ func (s *Service) removePermissionsByType(ctx context.Context, ID, pType string)
 		break
 	}
 
+	if len(permissions) == 0 {
+		return
+	}
+
 	if err := s.ofga.DeleteTuples(ctx, permissions...); err != nil {
+		s.logger.Error(err.Error())
+	}
+}
+
+func (s *Service) removeAssignees(ctx context.Context, ID string) {
+	ctx, span := s.tracer.Start(ctx, "roles.Service.removeAssignees")
+	defer span.End()
+
+	cToken := ""
+	assignees := make([]ofga.Tuple, 0)
+	for {
+		r, err := s.ofga.ReadTuples(ctx, "", ASSIGNEE_RELATION, fmt.Sprintf("role:%s", ID), cToken)
+
+		if err != nil {
+			s.logger.Errorf("error when retrieving tuples for %s role:%s", ASSIGNEE_RELATION, ID)
+			return
+		}
+
+		for _, t := range r.Tuples {
+			assignees = append(assignees, *ofga.NewTuple(t.Key.User, t.Key.Relation, t.Key.Object))
+		}
+
+		// if there are more pages, keep going with the loop
+		if cToken = r.ContinuationToken; cToken != "" {
+			continue
+		}
+
+		// TODO @shipperizer understand if better breaking at every cycle or reverting if clause
+		break
+	}
+
+	if len(assignees) == 0 {
+		return
+	}
+
+	if err := s.ofga.DeleteTuples(ctx, assignees...); err != nil {
 		s.logger.Error(err.Error())
 	}
 }
@@ -352,6 +401,12 @@ func (s *Service) listPermissionsFunc(ctx context.Context, roleID, ofgaType, cTo
 func (s *Service) removePermissionsFunc(ctx context.Context, roleID, ofgaType string) func() {
 	return func() {
 		s.removePermissionsByType(ctx, roleID, ofgaType)
+	}
+}
+
+func (s *Service) removeAssigneesFunc(ctx context.Context, roleID string) func() {
+	return func() {
+		s.removeAssignees(ctx, roleID)
 	}
 }
